+++ Tools/scripts/stable_abi.py
@@ -21,9 +21,7 @@ EXCLUDED_HEADERS = {
     "genobject.h",
     "longintrepr.h",
     "parsetok.h",
-    "pyarena.h",
     "pyatomic.h",
-    "pyctype.h",
     "pydebug.h",
     "pytime.h",
     "symtable.h",
